+++ programs/nft_voter/src/instructions/initialize_nft_voter_v0.rs
@@ -1,7 +1,8 @@
-use crate::state::*;
 use anchor_lang::prelude::*;
 use anchor_spl::token::Mint;
 use nft_proxy::state::ProxyConfigV0;
+
+use crate::state::*;
 
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Default)]
 pub struct InitializeNftVoterArgsV0 {
@@ -19,7 +20,7 @@ pub struct InitializeNftVoterV0<'info> {
   #[account(
     init,
     payer = payer,
-    space = 8 + 60 + std::mem::size_of::<NftVoterV0>(),
+    space = 8 + 60 + std::mem::size_of::<NftVoterV0>() + 4 + args.name.len(),
     seeds = [b"nft_voter", args.name.as_bytes()],
     bump
   )]
@@ -36,7 +37,7 @@ pub fn handler(ctx: Context<InitializeNftVoterV0>, args: InitializeNftVoterArgsV
     .map(|k| k.key())
     .unwrap_or_default();
   ctx.accounts.nft_voter.set_inner(NftVoterV0 {
-    bump_seed: ctx.bumps["nft_voter"],
+    bump_seed: ctx.bumps.nft_voter,
     name: args.name,
     authority: args.authority,
     collection: ctx.accounts.collection.key(),