+++ programs/nft_voter/src/instructions/proxied_relinquish_vote_v0.rs
@@ -1,10 +1,12 @@
-use crate::{error::ErrorCode, metaplex::MetadataAccount, RelinquishVoteArgsV0};
 use anchor_lang::prelude::*;
-use anchor_spl::token::Mint;
+use anchor_spl::{
+  metadata::{mpl_token_metadata, MetadataAccount},
+  token::Mint,
+};
 use nft_proxy::state::ProxyAssignmentV0;
 use proposal::{ProposalConfigV0, ProposalV0};
 
-use crate::{nft_voter_seeds, state::*};
+use crate::{error::ErrorCode, nft_voter_seeds, state::*, RelinquishVoteArgsV0};
 
 #[derive(Accounts)]
 pub struct ProxiedRelinquishVoteV0<'info> {
@@ -25,7 +27,7 @@ pub struct ProxiedRelinquishVoteV0<'info> {
   pub mint: Box<Account<'info, Mint>>,
   #[account(
     seeds = ["metadata".as_bytes(), MetadataAccount::owner().as_ref(), mint.key().as_ref()],
-    seeds::program = MetadataAccount::owner(),
+    seeds::program = mpl_token_metadata::ID,
     bump,
     constraint = metadata.collection.as_ref().map(|col| col.verified && col.key == nft_voter.collection).unwrap_or_else(|| false)
   )]
@@ -35,6 +37,7 @@ pub struct ProxiedRelinquishVoteV0<'info> {
     constraint = proxy_assignment.proxy_config == nft_voter.proxy_config,
     constraint = proxy_assignment.index <= marker.proxy_index,
     constraint = proxy_assignment.expiration_time > Clock::get().unwrap().unix_timestamp,
+    constraint = proxy_assignment.asset == mint.key(),
   )]
   pub proxy_assignment: Box<Account<'info, ProxyAssignmentV0>>,
   #[account(