+++ programs/nft_voter/src/instructions/proxied_vote_v0.rs
@@ -1,10 +1,12 @@
-use crate::{error::ErrorCode, metaplex::MetadataAccount, VoteArgsV0};
 use anchor_lang::prelude::*;
-use anchor_spl::token::Mint;
+use anchor_spl::{
+  metadata::{mpl_token_metadata, MetadataAccount},
+  token::Mint,
+};
 use nft_proxy::state::ProxyAssignmentV0;
 use proposal::{ProposalConfigV0, ProposalV0};
 
-use crate::{nft_voter_seeds, state::*};
+use crate::{error::ErrorCode, nft_voter_seeds, state::*, VoteArgsV0};
 
 #[derive(Accounts)]
 pub struct ProxyVoteV0<'info> {
@@ -13,7 +15,7 @@ pub struct ProxyVoteV0<'info> {
   #[account(
     init_if_needed,
     payer = payer,
-    space = 8 + 60 + std::mem::size_of::<VoteMarkerV0>(),
+    space = 8 + 60 + std::mem::size_of::<VoteMarkerV0>() + 2 * proposal.choices.len(),
     seeds = [b"marker", nft_voter.key().as_ref(), mint.key().as_ref(), proposal.key().as_ref()],
     bump
   )]
@@ -24,14 +26,15 @@ pub struct ProxyVoteV0<'info> {
     // only the current or earlier delegates can change vote. Or if proposal not set, this was an `init` for the marker
     constraint = proxy_assignment.index <= marker.proxy_index || marker.proposal == Pubkey::default(),
     constraint = proxy_assignment.expiration_time > Clock::get().unwrap().unix_timestamp,
+    constraint = proxy_assignment.asset == mint.key(),
   )]
   pub proxy_assignment: Box<Account<'info, ProxyAssignmentV0>>,
   pub nft_voter: Box<Account<'info, NftVoterV0>>,
   pub voter: Signer<'info>,
   pub mint: Box<Account<'info, Mint>>,
   #[account(
-    seeds = ["metadata".as_bytes(), MetadataAccount::owner().as_ref(), mint.key().as_ref()],
-    seeds::program = MetadataAccount::owner(),
+    seeds = ["metadata".as_bytes(), mpl_token_metadata::ID.as_ref(), mint.key().as_ref()],
+    seeds::program = mpl_token_metadata::ID,
     bump,
     constraint = metadata.collection.as_ref().map(|col| col.verified && col.key == nft_voter.collection).unwrap_or_else(|| false)
   )]
@@ -67,7 +70,7 @@ pub fn handler(ctx: Context<ProxyVoteV0>, args: VoteArgsV0) -> Result<()> {
     marker.rent_refund = ctx.accounts.payer.key();
   }
   marker.proposal = ctx.accounts.proposal.key();
-  marker.bump_seed = ctx.bumps["marker"];
+  marker.bump_seed = ctx.bumps.marker;
   marker.voter = ctx.accounts.voter.key();
   marker.nft_voter = ctx.accounts.nft_voter.key();
   marker.mint = ctx.accounts.mint.key();