+++ programs/state_controller/src/state.rs
@@ -42,6 +42,19 @@ pub enum ResolutionNode {
   },
   And,
   Or,
+  Not {
+    choice_name: String,
+  },
+  TotalWeight {
+    weight_threshold: u128,
+  },
+  /// The choice has this percentage (i32 / PERCENTAGE_DIVISOR)
+  /// This is different than ChoicePercentage as it operates based off of the
+  /// current set of choices on the stack, not all of the proposals choices.
+  /// This is useful for use cases like Abstain, where you need a % excluding abstain.
+  ChoicePercentageOfCurrent {
+    percentage: i32,
+  },
 }
 
 impl Default for ResolutionNode {
@@ -60,10 +73,13 @@ impl ResolutionNode {
       ResolutionNode::OffsetFromStartTs { .. } => 8,
       ResolutionNode::ChoiceVoteWeight { .. } => 16,
       ResolutionNode::ChoicePercentage { .. } => 4,
+      ResolutionNode::ChoicePercentageOfCurrent { .. } => 4,
       ResolutionNode::Top { .. } => 2,
-      ResolutionNode::And => 0,
-      ResolutionNode::Or => 0,
+      ResolutionNode::And => 1,
+      ResolutionNode::Or => 1,
       ResolutionNode::NumResolved { .. } => 4,
+      ResolutionNode::Not { choice_name } => 4 + choice_name.len(),
+      ResolutionNode::TotalWeight { .. } => 16,
     }
   }
 
@@ -81,6 +97,11 @@ impl ResolutionNode {
         Err(error!(ErrorCode::InvalidOffset))
       }
       ResolutionNode::ChoicePercentage { percentage }
+        if *percentage < 0 || *percentage > PERCENTAGE_DIVISOR =>
+      {
+        Err(error!(ErrorCode::InvalidPercentage))
+      }
+      ResolutionNode::ChoicePercentageOfCurrent { percentage }
         if *percentage < 0 || *percentage > PERCENTAGE_DIVISOR =>
       {
         Err(error!(ErrorCode::InvalidPercentage))
@@ -216,17 +237,69 @@ impl ResolutionStrategy {
           );
           stack.push(ret)
         }
-        ResolutionNode::Top { n } => {
-          let mut vec = proposal.choices.iter().enumerate().collect::<Vec<_>>();
+        ResolutionNode::ChoicePercentageOfCurrent { percentage } => {
+          let remaining_choices = stack
+            .last()
+            .and_then(|i| i.clone())
+            .unwrap_or((0..proposal.choices.len() as u16).collect::<Vec<u16>>());
 
-          vec.sort_by(|(_, a), (_, b)| b.weight.cmp(&a.weight));
+          let total_weight = remaining_choices
+            .iter()
+            .map(|choice| proposal.choices[*choice as usize].weight)
+            .sum::<u128>();
+
+          let threshold = total_weight
+            .checked_mul(*percentage as u128)
+            .unwrap()
+            .checked_div(PERCENTAGE_DIVISOR as u128)
+            .map(|result| {
+              let remainder = total_weight
+                .checked_mul(*percentage as u128)
+                .unwrap()
+                .checked_rem(PERCENTAGE_DIVISOR as u128)
+                .unwrap();
+              result
+                .checked_add(if remainder > 0 { 1 } else { 0 })
+                .unwrap()
+            })
+            .unwrap();
+
+          let ret: Option<Vec<u16>> = Some(
+            remaining_choices
+              .iter()
+              .enumerate()
+              .flat_map(|(index, choice)| {
+                if threshold == 0 {
+                  None
+                } else if proposal.choices[*choice as usize].weight >= threshold {
+                  Some(index as u16)
+                } else {
+                  None
+                }
+              })
+              .collect(),
+          );
+
+          stack.push(ret)
+        }
+        ResolutionNode::Top { n } => {
+          let mut remaining_choices = stack
+            .last()
+            .and_then(|i| i.clone())
+            .unwrap_or((0..proposal.choices.len() as u16).collect::<Vec<u16>>())
+            .clone();
+
+          remaining_choices.sort_by(|a, b| {
+            proposal.choices[*b as usize]
+              .weight
+              .cmp(&proposal.choices[*a as usize].weight)
+          });
 
           stack.push(Some(
-            vec
-              .iter()
-              .map(|(index, _)| *index as u16)
+            remaining_choices
+              .into_iter()
               .take(*n as usize)
-              .collect(),
+              .collect::<Vec<_>>(),
           ))
         }
         ResolutionNode::And => {
@@ -258,6 +331,35 @@ impl ResolutionStrategy {
           match curr {
             Some(vec) if vec.len() >= *n as usize => stack.push(Some(vec.clone())),
             _ => stack.push(None),
+          }
+        }
+        ResolutionNode::Not { choice_name } => {
+          let vec = proposal
+            .choices
+            .iter()
+            .enumerate()
+            .filter(|(_, choice)| choice.name != *choice_name)
+            .collect::<Vec<_>>();
+
+          stack.push(Some(vec.iter().map(|(index, _)| *index as u16).collect()))
+        }
+        ResolutionNode::TotalWeight { weight_threshold } => {
+          let total_weight = proposal
+            .choices
+            .iter()
+            .map(|choice| choice.weight)
+            .sum::<u128>();
+          if total_weight >= *weight_threshold {
+            stack.push(Some(
+              proposal
+                .choices
+                .iter()
+                .enumerate()
+                .map(|(index, _)| index as u16)
+                .collect(),
+            ))
+          } else {
+            stack.push(None)
           }
         }
       }