+++ programs/token_voter/src/instructions/withdraw_v0.rs
@@ -1,10 +1,11 @@
-use crate::metaplex::{burn, Burn, Metadata};
-use crate::receipt_seeds;
-use crate::state::*;
 use anchor_lang::prelude::*;
-use anchor_spl::associated_token::AssociatedToken;
-use anchor_spl::token::{self, CloseAccount, Transfer};
-use anchor_spl::token::{Mint, Token, TokenAccount};
+use anchor_spl::{
+  associated_token::AssociatedToken,
+  metadata::{burn_nft, mpl_token_metadata, BurnNft, Metadata},
+  token::{self, CloseAccount, Mint, Token, TokenAccount, Transfer},
+};
+
+use crate::{receipt_seeds, state::*};
 
 #[derive(Accounts)]
 pub struct WithdrawV0<'info> {
@@ -39,8 +40,8 @@ pub struct WithdrawV0<'info> {
 
   #[account(
     mut,
-    seeds = ["metadata".as_bytes(), token_metadata_program.key().as_ref(), mint.key().as_ref()],
-    seeds::program = token_metadata_program.key(),
+    seeds = ["metadata".as_bytes(),  mpl_token_metadata::ID.as_ref(), mint.key().as_ref()],
+    seeds::program = mpl_token_metadata::ID,
     bump,
   )]
   /// CHECK: Checked by cpi
@@ -48,8 +49,8 @@ pub struct WithdrawV0<'info> {
   /// CHECK: Handled by cpi
   #[account(
     mut,
-    seeds = ["metadata".as_bytes(), token_metadata_program.key().as_ref(), mint.key().as_ref(), "edition".as_bytes()],
-    seeds::program = token_metadata_program.key(),
+    seeds = ["metadata".as_bytes(),  mpl_token_metadata::ID.as_ref(), mint.key().as_ref(), "edition".as_bytes()],
+    seeds::program = mpl_token_metadata::ID,
     bump,
   )]
   pub master_edition: UncheckedAccount<'info>,
@@ -62,7 +63,6 @@ pub struct WithdrawV0<'info> {
 
   #[account(
     mut,
-    close = refund,
     associated_token::authority = receipt,
     associated_token::mint = deposit_mint,
   )]
@@ -88,18 +88,17 @@ pub struct WithdrawV0<'info> {
   pub token_program: Program<'info, Token>,
   pub associated_token_program: Program<'info, AssociatedToken>,
   pub token_metadata_program: Program<'info, Metadata>,
-  pub rent: Sysvar<'info, Rent>,
 }
 
 impl<'info> WithdrawV0<'info> {
-  fn burn_nft_ctx(&self) -> CpiContext<'_, '_, '_, 'info, Burn<'info>> {
-    let cpi_accounts = Burn {
+  fn burn_nft_ctx(&self) -> CpiContext<'_, '_, '_, 'info, BurnNft<'info>> {
+    let cpi_accounts = BurnNft {
       metadata: self.metadata.to_account_info(),
       owner: self.owner.to_account_info(),
       mint: self.mint.to_account_info(),
-      token_account: self.receipt_token_account.to_account_info(),
-      master_edition_account: self.master_edition.to_account_info(),
-      collection: self.collection_metadata.to_account_info(),
+      token: self.receipt_token_account.to_account_info(),
+      edition: self.master_edition.to_account_info(),
+      spl_token: self.token_program.to_account_info(),
     };
     CpiContext::new(self.token_metadata_program.to_account_info(), cpi_accounts)
   }
@@ -126,7 +125,13 @@ impl<'info> WithdrawV0<'info> {
 pub fn handler(ctx: Context<WithdrawV0>) -> Result<()> {
   let signer_seeds: &[&[&[u8]]] = &[receipt_seeds!(ctx.accounts.receipt)];
 
-  burn(ctx.accounts.burn_nft_ctx())?;
+  burn_nft(
+    ctx
+      .accounts
+      .burn_nft_ctx()
+      .with_remaining_accounts(vec![ctx.accounts.collection_metadata.to_account_info()]),
+    Some(ctx.accounts.collection_metadata.key()),
+  )?;
   token::transfer(
     ctx
       .accounts