+++ tests/organization-wallet.ts
@@ -82,6 +82,7 @@ describe("organization wallet", () => {
           voteController: me,
           stateController: me,
           onVoteHook: PublicKey.default,
+          authority: me,
         })
         .rpcAndKeys());
 
@@ -107,7 +108,7 @@ describe("organization wallet", () => {
           name: "My Wallet",
           proposalConfigs: [proposalConfig!],
         })
-        .accounts({
+        .accountsPartial({
           organization,
         })
         .rpcAndKeys({ skipPreflight: true });
@@ -139,7 +140,7 @@ describe("organization wallet", () => {
             name: "My Wallet",
             proposalConfigs: [proposalConfig!],
           })
-          .accounts({
+          .accountsPartial({
             organization,
           })
           .rpcAndKeys({ skipPreflight: true }));
@@ -162,7 +163,7 @@ describe("organization wallet", () => {
             ],
             tags: ["test", "tags"],
           })
-          .accounts({ organization })
+          .accountsPartial({ organization })
           .rpcAndKeys({ skipPreflight: true }));
       });
 
@@ -208,7 +209,7 @@ describe("organization wallet", () => {
             disableExecutionOffset: 60 * 60 * 24 * 7,
           })
           .remainingAccounts(remainingAccounts)
-          .accounts({
+          .accountsPartial({
             proposal,
             organizationWallet,
           })
@@ -225,7 +226,7 @@ describe("organization wallet", () => {
               },
             },
           })
-          .accounts({
+          .accountsPartial({
             proposal,
           })
           .rpc({ skipPreflight: true });
@@ -239,7 +240,7 @@ describe("organization wallet", () => {
       });
 
       describe("update_organization_wallet_v0", () => {
-        let otherName;
+        let otherName: string;
         let otherProposalConfig: PublicKey | undefined;
 
         beforeEach(async () => {
@@ -253,6 +254,7 @@ describe("organization wallet", () => {
               voteController: me,
               stateController: me,
               onVoteHook: PublicKey.default,
+              authority: me,
             })
             .rpcAndKeys({ skipPreflight: true }));
         });
@@ -263,7 +265,7 @@ describe("organization wallet", () => {
               name: otherName,
               proposalConfigs: null,
             })
-            .accounts({
+            .accountsPartial({
               organizationWallet,
               organization,
               authority: me,
@@ -286,7 +288,7 @@ describe("organization wallet", () => {
               name: null,
               proposalConfigs: [otherProposalConfig!],
             })
-            .accounts({
+            .accountsPartial({
               organizationWallet,
               organization,
               authority: me,
@@ -313,7 +315,7 @@ describe("organization wallet", () => {
               name: null,
               proposalConfigs,
             })
-            .accounts({
+            .accountsPartial({
               organizationWallet,
               organization,
               authority: me,
@@ -337,7 +339,7 @@ describe("organization wallet", () => {
               proposalProgram: null,
               uri: null,
             })
-            .accounts({
+            .accountsPartial({
               organization,
               authority: me,
             })
@@ -351,13 +353,13 @@ describe("organization wallet", () => {
                 name: otherName,
                 proposalConfigs: [otherProposalConfig!],
               })
-              .accounts({
+              .accountsPartial({
                 organizationWallet,
                 organization,
                 authority: me,
               })
               .simulate();
-          } catch (err) {
+          } catch (err: any) {
             logs = err.simulationResponse?.logs;
           }
 
@@ -387,13 +389,13 @@ describe("organization wallet", () => {
                 name: otherName,
                 proposalConfigs: [otherProposalConfig!],
               })
-              .accounts({
+              .accountsPartial({
                 organizationWallet,
                 organization: otherOrganization,
                 authority: me,
               })
               .simulate();
-          } catch (err) {
+          } catch (err: any) {
             logs = err.simulationResponse?.logs;
           }