+++ tests/proposal.ts
@@ -32,6 +32,7 @@ describe("proposal", () => {
         voteController,
         stateController,
         onVoteHook,
+        authority: me,
       })
       .rpcAndKeys();
 
@@ -40,7 +41,7 @@ describe("proposal", () => {
     expect(acct.voteController.toBase58()).to.eq(voteController.toBase58());
     expect(acct.stateController.toBase58()).to.eq(stateController.toBase58());
     expect(acct.onVoteHook.toBase58()).to.eq(onVoteHook.toBase58());
-    expect(acct.authority.toBase58()).to.eq(PublicKey.default.toBase58());
+    expect(acct.authority.toBase58()).to.eq(me.toBase58());
   });
 
   it("Creates a proposal config with authority", async () => {
@@ -95,7 +96,7 @@ describe("proposal", () => {
           onVoteHook,
           authority,
         })
-        .accounts({ proposalConfig })
+        .accountsPartial({ proposalConfig })
         .rpc();
 
       const acct = await program.account.proposalConfigV0.fetch(
@@ -124,7 +125,7 @@ describe("proposal", () => {
 
       await program.methods
         .updateProposalConfigV0(args as any)
-        .accounts({ proposalConfig })
+        .accountsPartial({ proposalConfig })
         .rpc();
 
       const acct = await program.account.proposalConfigV0.fetch(
@@ -156,7 +157,7 @@ describe("proposal", () => {
           onVoteHook: null,
           authority,
         })
-        .accounts({ proposalConfig })
+        .accountsPartial({ proposalConfig })
         .rpc();
 
       try {
@@ -167,11 +168,11 @@ describe("proposal", () => {
             onVoteHook: PublicKey.default,
             authority: me,
           })
-          .accounts({
+          .accountsPartial({
             proposalConfig,
           })
           .simulate();
-      } catch (err) {
+      } catch (err: any) {
         logs = err.simulationResponse?.logs;
       }
 
@@ -201,7 +202,7 @@ describe("proposal", () => {
           ],
           tags: ["test", "tags"],
         })
-        .accounts({ proposalConfig })
+        .accountsPartial({ proposalConfig })
         .rpcAndKeys();
 
       const acct = await program.account.proposalV0.fetch(proposal!);
@@ -238,7 +239,7 @@ describe("proposal", () => {
               ],
               tags: ["test", "tags"],
             })
-            .accounts({ proposalConfig })
+            .accountsPartial({ proposalConfig })
             .rpcAndKeys()
         ).pubkeys.proposal!;
       });
@@ -252,7 +253,7 @@ describe("proposal", () => {
               },
             },
           })
-          .accounts({ proposal })
+          .accountsPartial({ proposal })
           .rpc();
         await program.methods
           .voteV0({
@@ -260,7 +261,7 @@ describe("proposal", () => {
             weight: new anchor.BN(2),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         let acct = await program.account.proposalV0.fetch(proposal);
@@ -272,7 +273,7 @@ describe("proposal", () => {
             weight: new anchor.BN(1),
             removeVote: true,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc();
 
         acct = await program.account.proposalV0.fetch(proposal);
@@ -284,7 +285,7 @@ describe("proposal", () => {
           .updateStateV0({
             newState: { custom: { name: "hello", bin: Buffer.from([]) } },
           })
-          .accounts({ proposal })
+          .accountsPartial({ proposal })
           .rpc({ skipPreflight: true });
 
         const acct = await program.account.proposalV0.fetch(proposal);