+++ tests/state-controller.ts
@@ -62,6 +62,7 @@ describe("state-controller", () => {
           voteController: me,
           stateController: resolutionSettings!,
           onVoteHook: PROGRAM_ID,
+          authority: me,
         })
         .rpcAndKeys({ skipPreflight: true }));
       ({
@@ -81,17 +82,25 @@ describe("state-controller", () => {
               name: "No",
               uri: null,
             },
+            {
+              name: "Abstain",
+              uri: null,
+            },
           ],
           tags: ["test", "tags"],
         })
-        .accounts({ proposalConfig })
+        .accountsPartial({ proposalConfig })
         .rpcAndKeys({ skipPreflight: true }));
 
       await program.methods
         .updateStateV0({
-          newState: { voting: {} },
+          newState: {
+            voting: {
+              startTs: new anchor.BN(Date.now() / 1000),
+            },
+          },
         })
-        .accounts({ proposal })
+        .accountsPartial({ proposal })
         .rpc();
     });
 
@@ -101,7 +110,7 @@ describe("state-controller", () => {
       });
 
       it("resolves to the choice selected", async () => {
-        await program.methods.resolveV0().accounts({ proposal }).rpc();
+        await program.methods.resolveV0().accountsPartial({ proposal }).rpc();
 
         const acct = await proposalProgram.account.proposalV0.fetch(proposal!);
         expect(acct.state.resolved?.choices).to.deep.eq([1]);
@@ -114,10 +123,10 @@ describe("state-controller", () => {
       });
 
       it("resolves to all choices", async () => {
-        await program.methods.resolveV0().accounts({ proposal }).rpc();
+        await program.methods.resolveV0().accountsPartial({ proposal }).rpc();
 
         const acct = await proposalProgram.account.proposalV0.fetch(proposal!);
-        expect(acct.state.resolved?.choices).to.deep.eq([0, 1]);
+        expect(acct.state.resolved?.choices).to.deep.eq([0, 1, 2]);
       });
     });
 
@@ -127,7 +136,7 @@ describe("state-controller", () => {
       });
 
       it("not resolve", async () => {
-        await program.methods.resolveV0().accounts({ proposal }).rpc();
+        await program.methods.resolveV0().accountsPartial({ proposal }).rpc();
 
         const acct = await proposalProgram.account.proposalV0.fetch(proposal!);
         expect(Boolean(acct.state.voting)).to.be.true;
@@ -141,10 +150,10 @@ describe("state-controller", () => {
 
       it("resolves to all choices", async () => {
         await sleep(3000);
-        await program.methods.resolveV0().accounts({ proposal }).rpc();
+        await program.methods.resolveV0().accountsPartial({ proposal }).rpc();
 
         const acct = await proposalProgram.account.proposalV0.fetch(proposal!);
-        expect(acct.state.resolved?.choices).to.deep.eq([0, 1]);
+        expect(acct.state.resolved?.choices).to.deep.eq([0, 1, 2]);
       });
     });
 
@@ -165,7 +174,7 @@ describe("state-controller", () => {
             weight: new anchor.BN(1),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         const acct = await proposalProgram.account.proposalV0.fetch(proposal!);
@@ -187,7 +196,7 @@ describe("state-controller", () => {
             weight: new anchor.BN(1),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         const acct = await proposalProgram.account.proposalV0.fetch(proposal!);
@@ -218,7 +227,7 @@ describe("state-controller", () => {
             weight: new anchor.BN(1),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         const acct = await proposalProgram.account.proposalV0.fetch(proposal!);
@@ -249,7 +258,7 @@ describe("state-controller", () => {
             weight: new anchor.BN(5),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         let acct = await proposalProgram.account.proposalV0.fetch(proposal!);
@@ -261,12 +270,12 @@ describe("state-controller", () => {
             weight: new anchor.BN(3),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         await sleep(10000);
 
-        await program.methods.resolveV0().accounts({ proposal }).rpc();
+        await program.methods.resolveV0().accountsPartial({ proposal }).rpc();
 
         acct = await proposalProgram.account.proposalV0.fetch(proposal!);
         expect(acct.state.resolved?.choices).to.deep.eq([1]);
@@ -279,7 +288,7 @@ describe("state-controller", () => {
             weight: new anchor.BN(2),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         let acct = await proposalProgram.account.proposalV0.fetch(proposal!);
@@ -291,11 +300,11 @@ describe("state-controller", () => {
             weight: new anchor.BN(3),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         await sleep(10000);
-        await program.methods.resolveV0().accounts({ proposal }).rpc();
+        await program.methods.resolveV0().accountsPartial({ proposal }).rpc();
 
         acct = await proposalProgram.account.proposalV0.fetch(proposal!);
         expect(acct.state.resolved?.choices).to.deep.eq([]);
@@ -322,10 +331,10 @@ describe("state-controller", () => {
             weight: new anchor.BN(3),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
-        await program.methods.resolveV0().accounts({ proposal }).rpc();
+        await program.methods.resolveV0().accountsPartial({ proposal }).rpc();
 
         let acct = await proposalProgram.account.proposalV0.fetch(proposal!);
         expect(Boolean(acct.state.voting)).to.be.true;
@@ -336,7 +345,7 @@ describe("state-controller", () => {
             weight: new anchor.BN(5),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         acct = await proposalProgram.account.proposalV0.fetch(proposal!);
@@ -350,17 +359,17 @@ describe("state-controller", () => {
             weight: new anchor.BN(3),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
-        await program.methods.resolveV0().accounts({ proposal }).rpc();
+        await program.methods.resolveV0().accountsPartial({ proposal }).rpc();
 
         let acct = await proposalProgram.account.proposalV0.fetch(proposal!);
         expect(Boolean(acct.state.voting)).to.be.true;
 
         await sleep(6000);
 
-        await program.methods.resolveV0().accounts({ proposal }).rpc();
+        await program.methods.resolveV0().accountsPartial({ proposal }).rpc();
 
         acct = await proposalProgram.account.proposalV0.fetch(proposal!);
         expect(acct.state.resolved?.choices).to.deep.eq([]);
@@ -390,7 +399,7 @@ describe("state-controller", () => {
             weight: new anchor.BN(5),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         let acct = await proposalProgram.account.proposalV0.fetch(proposal!);
@@ -402,11 +411,63 @@ describe("state-controller", () => {
             weight: new anchor.BN(1),
             removeVote: false,
           })
-          .accounts({ proposal, voter: me })
+          .accountsPartial({ proposal, voter: me })
           .rpc({ skipPreflight: true });
 
         await sleep(10000);
-        await program.methods.resolveV0().accounts({ proposal }).rpc();
+        await program.methods.resolveV0().accountsPartial({ proposal }).rpc();
+
+        acct = await proposalProgram.account.proposalV0.fetch(proposal!);
+        expect(acct.state.resolved?.choices).to.deep.eq([1]);
+      });
+    });
+
+    describe("helium IOT flavor governance", () => {
+      before(async () => {
+        nodes = settings()
+          .and(
+            settings().offsetFromStartTs(new anchor.BN(5)),
+            settings().and(
+              settings().and(
+                settings().totalWeight(new anchor.BN(5)),
+                settings().and(
+                  settings().not("Abstain"),
+                  settings().choicePercentageOfCurrent(66.6)
+                )
+              ),
+              settings().top()
+            )
+          )
+          .build();
+      });
+
+      it("resolves to the max choice when there is enough weight", async () => {
+        await proposalProgram.methods
+          .voteV0({
+            choice: 1,
+            weight: new anchor.BN(2),
+            removeVote: false,
+          })
+          .accountsPartial({ proposal, voter: me })
+          .rpc({ skipPreflight: true });
+
+        let acct = await proposalProgram.account.proposalV0.fetch(proposal!);
+        expect(Boolean(acct.state.voting)).to.be.true;
+
+        await proposalProgram.methods
+          .voteV0({
+            choice: 2,
+            weight: new anchor.BN(5),
+            removeVote: false,
+          })
+          .accountsPartial({ proposal, voter: me })
+          .rpc({ skipPreflight: true });
+
+        await sleep(10000);
+        console.log(
+          "txid",
+          await program.methods.resolveV0().accountsPartial({ proposal }).rpc()
+        );
 
         acct = await proposalProgram.account.proposalV0.fetch(proposal!);
         expect(acct.state.resolved?.choices).to.deep.eq([1]);