+++ packages/organization-wallet-sdk/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.5](https://github.com/helium/modular-governance/compare/v0.0.15...v0.1.5) (2025-03-21)
+
+**Note:** Version bump only for package @helium/organization-wallet-sdk
+
+
+
+
+
+## [0.1.5-alpha.0](https://github.com/helium/modular-governance/compare/v0.0.15...v0.1.5-alpha.0) (2025-03-19)
+
+**Note:** Version bump only for package @helium/organization-wallet-sdk
+
+
+
+
+
+## [0.1.4](https://github.com/helium/modular-governance/compare/v0.0.15...v0.1.4) (2025-02-24)
+
+**Note:** Version bump only for package @helium/organization-wallet-sdk
+
+
+
+
+
+## [0.1.3](https://github.com/helium/modular-governance/compare/v0.0.15...v0.1.3) (2025-02-24)
+
+**Note:** Version bump only for package @helium/organization-wallet-sdk
+
+
+
+
+
+## [0.1.2](https://github.com/helium/modular-governance/compare/v0.0.15...v0.1.2) (2025-02-24)
+
+**Note:** Version bump only for package @helium/organization-wallet-sdk
+
+
+
+
+
+## [0.1.1](https://github.com/helium/modular-governance/compare/v0.0.15...v0.1.1) (2025-02-14)
+
+**Note:** Version bump only for package @helium/organization-wallet-sdk
+
+
+
+
+
+# [0.1.0](https://github.com/helium/modular-governance/compare/v0.0.15...v0.1.0) (2025-02-14)
+
+**Note:** Version bump only for package @helium/organization-wallet-sdk
+
+
+
+
+
+## [0.0.15](https://github.com/helium/modular-governance/compare/v0.0.14...v0.0.15) (2024-12-06)
+
+**Note:** Version bump only for package @helium/organization-wallet-sdk
+
+
+
+
+
+## [0.0.14](https://github.com/helium/modular-governance/compare/v0.0.13...v0.0.14) (2024-10-24)
+
+**Note:** Version bump only for package @helium/organization-wallet-sdk
+
+
+
+
+
 ## [0.0.13](https://github.com/helium/modular-governance/compare/v0.0.12...v0.0.13) (2024-09-06)
 
 **Note:** Version bump only for package @helium/organization-wallet-sdk